inspirational_quotes/requirements.txt
openai
requests
moviepy
python-dotenv
lumaai
opencv-python
pillow